refacotr: organize files
