--- lib/core/core.dart
@@ -1,1 +1,0 @@
-export './breakpoints.dart';
