+++ lib/pages/home/widgets/sections/top_section.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: prefer_const_constructors
 
-import 'package:app_responsive/core/core.dart';
+import 'package:app_responsive/common/core.dart';
 import 'package:app_responsive/pages/home/widgets/fields/fields.dart';
 import 'package:flutter/material.dart';
